Grades Find the number of D, C, B and A grades for the last test on informatics, where n students from a class have successfully passed the test.
The program gets number n as the first line of input and then gets the grades themselves, each on a new line.

The program must output four numbers in a single line: the number of D, C, B, and A grades (in that order), separated by blank space characters.

Sample Input:
14
B
A
A
A
A
B
D
D
A
B
C
B
B
A

Sample Output:
2 1 5 6